address Mingwei feedback

+++ hydroflow/src/builder/surface/filter.rs
@@ -51,7 +51,7 @@ where
     Prev: PullSurface + TrackPullDependencies,
     Func: FnMut(&Context<'_>, &Prev::ItemOut) -> bool,
 {
-    fn insert_dep(&self, e: &mut super::DirectedEdgeSet) -> u16 {
+    fn insert_dep(&self, e: &mut super::DirectedEdgeSet) -> usize {
         let my_id = e.add_node("Filter".to_string());
         let prev_id = self.prev.insert_dep(e);
         e.add_edge((prev_id, my_id));
@@ -81,7 +81,7 @@ where
     Prev: PushSurface + TrackPushDependencies,
     Func: FnMut(&Context<'_>, &Prev::ItemOut) -> bool,
 {
-    fn insert_dep(&self, e: &mut super::DirectedEdgeSet) -> u16 {
+    fn insert_dep(&self, e: &mut super::DirectedEdgeSet) -> usize {
         let my_id = e.add_node("Filter".to_string());
         let prev_id = self.prev.insert_dep(e);
         e.add_edge((prev_id, my_id));
@@ -110,7 +110,7 @@ where
     Next: PushSurfaceReversed + TrackPushDependencies,
     Func: FnMut(&Context<'_>, &Next::ItemIn) -> bool,
 {
-    fn insert_dep(&self, e: &mut super::DirectedEdgeSet) -> u16 {
+    fn insert_dep(&self, e: &mut super::DirectedEdgeSet) -> usize {
         let my_id = e.add_node("Filter".to_string());
         let next_id = self.next.insert_dep(e);
         e.add_edge((my_id, next_id));
