+++ hydroflow/src/builder/surface/push_handoff.rs
@@ -30,7 +30,7 @@ impl<Hof, In> TrackPushDependencies for HandoffPushSurfaceReversed<Hof, In>
 where
     Hof: Handoff + CanReceive<In>,
 {
-    fn insert_dep(&self, e: &mut super::DirectedEdgeSet) -> u16 {
+    fn insert_dep(&self, e: &mut super::DirectedEdgeSet) -> usize {
         let my_id = e.add_node(format!("Handoff_{}", self.port.handoff_id));
         my_id
     }
